Refactor and remove the initialize function
Shout out to Stylianos Gakis from the Android Study Group Slack for
suggesting that MainViewModel#initialize was a strange pattern. I seem
to have misunderstood app architecutre guidelines from Google. You can
use an idempotent initialize function if kicking off work once the view
is initialized. However, if we use ApolloClient's Flow-based API we can
instead map the Flow to the UI state in a couple of simple functions.

Another shout out to Now In Android's codebase for the Result sealed
class that we use to catch exceptions from ApolloClient.