Use GitHubRepositoryRepository in view model
This commit introduces the data layer to the view model and updates
all call sites. We also introduce a network package for the network
module and the auth interceptor to keep things organized.

This commit also includes the missing test for the repository
implementation that was left out of the last commit.